requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.1.0
    # via kombu
asgiref==3.5.0
    # via django
billiard==3.6.4.0
    # via celery
celery==5.2.3
    # via -r requirements.in
click==8.0.4
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.4
    # via click
coverage==6.3.2
    # via -r requirements.in
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
django==4.0.3
    # via
    #   -r requirements.in
    #   django-import-export
    #   djangorestframework
django-import-export==2.7.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
flake8==4.0.1
    # via -r requirements.in
kombu==5.2.4
    # via celery
lml==0.1.0
    # via pyexcel-io
markuppy==1.14
    # via tablib
mccabe==0.6.1
    # via flake8
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via
    #   pyexcel-xlsx
    #   tablib
pep517==0.12.0
    # via pip-tools
pip-tools==6.5.1
    # via -r requirements.in
prompt-toolkit==3.0.28
    # via click-repl
pycodestyle==2.8.0
    # via flake8
pyexcel-io==0.6.6
    # via pyexcel-xlsx
pyexcel-xlsx==0.6.0
    # via -r requirements.in
pyflakes==2.4.0
    # via flake8
pytz==2022.1
    # via
    #   celery
    #   djangorestframework
pyyaml==6.0
    # via tablib
six==1.16.0
    # via click-repl
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.2.0
    # via django-import-export
tomli==2.0.1
    # via pep517
tzdata==2022.1
    # via django
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.1
    # via pip-tools
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
